\name{circosR-brain-package}
\alias{circosR-brain-package}
\alias{circosR-brain}
\docType{package}
\title{
\packageTitle{circosR-brain}
}
\description{
\packageDescription{circosR-brain}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{circosR-brain}
\packageIndices{circosR-brain}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{circosR-brain}

Maintainer: \packageMaintainer{circosR-brain}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
